refactor(api): migrate GET /api/admin/tags/{id}/recently-used
