--- api/lib/application/tags/index.js
@@ -1,41 +1,0 @@
-import Joi from 'joi';
-
-import { securityPreHandlers } from '../../../src/shared/application/security-pre-handlers.js';
-import { identifiersType } from '../../../src/shared/domain/types/identifiers-type.js';
-import { tagController } from './tag-controller.js';
-
-const register = async function (server) {
-  server.route([
-    {
-      method: 'GET',
-      path: '/api/admin/tags/{id}/recently-used',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        handler: tagController.getRecentlyUsedTags,
-        validate: {
-          params: Joi.object({
-            id: identifiersType.tagId,
-          }),
-        },
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
-            '- Renvoie les 10 tags les plus utilisés par rapport au tag sélectionné',
-        ],
-        tags: ['api', 'tags'],
-      },
-    },
-  ]);
-};
-
-const name = 'tags-api';
-export { name, register };
