--- api/lib/application/tags/tag-controller.js
@@ -1,12 +1,0 @@
-import * as tagSerializer from '../../../src/organizational-entities/infrastructure/serializers/jsonapi/tag-serializer.js';
-import { usecases } from '../../domain/usecases/index.js';
-
-const getRecentlyUsedTags = async function (request, h, dependencies = { tagSerializer }) {
-  const tagId = request.params.id;
-  const recentlyUsedTags = await usecases.getRecentlyUsedTags({ tagId });
-  return dependencies.tagSerializer.serialize(recentlyUsedTags);
-};
-
-const tagController = { getRecentlyUsedTags };
-
-export { tagController };
