+++ api/lib/routes.js
@@ -9,7 +9,6 @@ import * as memberships from './application/memberships/index.js';
 import * as organizations from './application/organizations/index.js';
 import * as passwords from './application/passwords/index.js';
 import * as scoOrganizationLearners from './application/sco-organization-learners/index.js';
-import * as tags from './application/tags/index.js';
 import * as targetProfiles from './application/target-profiles/index.js';
 import * as users from './application/users/index.js';
 
@@ -24,7 +23,6 @@ const routes = [
   organizations,
   passwords,
   scoOrganizationLearners,
-  tags,
   targetProfiles,
   frameworks,
   users,
